Fix dependency cycle and format/lint
